Put each type into its own file
